feature: remvoed all components except for multiuse ones
