+++ src/app.postcss
@@ -121,4 +121,12 @@
 			}
 		}
 	}
+
+	.card {
+		@apply w-64 h-36 flex items-end justify-start text-left rounded-b-md rounded-tl-md mr-4 mb-4;
+
+		h2 {
+			@apply text-2xl p-2;
+		}
+	}
 }
